requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
affine=2.4.0=pyhd8ed1ab_0
aom=3.5.0=h63175ca_0
attrs=23.1.0=pyh71513ae_1
blas=2.116=mkl
blas-devel=3.9.0=16_win64_mkl
blosc=1.21.4=hdccc3a2_0
boost-cpp=1.78.0=h9f4b32c_3
brotli=1.0.9=hcfcfb64_8
brotli-bin=1.0.9=hcfcfb64_8
bzip2=1.0.8=h8ffe710_4
c-blosc2=2.9.2=h183a6f4_0
ca-certificates=2023.5.7=h56e8100_0
cairo=1.16.0=hd694305_1014
certifi=2023.5.7=pyhd8ed1ab_0
cfitsio=4.2.0=h9ebe7e4_0
charls=2.4.2=h1537add_0
charset-normalizer=3.1.0=pyhd8ed1ab_0
click=8.1.3=win_pyhd8ed1ab_2
click-plugins=1.1.1=py_0
cligj=0.7.2=pyhd8ed1ab_1
cloudpickle=2.2.1=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.0.7=py311h005e61a_0
cuda-cccl=12.1.109=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.1.105=0
cuda-runtime=11.8.0=0
curl=8.1.1=h68f0423_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.12.0=py311ha68e1ae_1
dask-core=2023.5.1=pyhd8ed1ab_0
dav1d=1.2.0=hcfcfb64_0
expat=2.5.0=h63175ca_1
filelock=3.12.0=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=hbde0cde_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
freexl=1.0.6=h67ca5e6_1
fsspec=2023.5.0=pyh1a96a4e_0
geos=3.11.1=h1537add_0
geotiff=1.7.1=hc256dc0_6
gettext=0.21.1=h5728263_0
giflib=5.2.1=h64bf75a_3
h5py=3.8.0=pypi_0
hdf4=4.2.15=h1b1b6ef_5
hdf5=1.12.2=nompi_h57737ce_101
icu=70.1=h0e60522_0
idna=3.4=pyhd8ed1ab_0
imagecodecs=2023.1.23=py311h8318e5d_0
imageio=2.28.1=pyh24c5eb1_0
importlib-metadata=6.6.0=pyha770c72_0
importlib_metadata=6.6.0=hd8ed1ab_0
intel-openmp=2023.1.0=h57928b3_46319
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h8ffe710_2
jxrlib=1.1=h8ffe710_2
kealib=1.5.0=h61be68b_0
kiwisolver=1.4.4=py311h005e61a_1
krb5=1.20.1=heb0366b_0
lazy_loader=0.2=pyhd8ed1ab_0
lcms2=2.15=ha5c8aab_0
lerc=4.0.0=h63175ca_0
libaec=1.0.6=h63175ca_1
libarchive=3.6.2=h27c7867_0
libavif=0.11.1=h5da7522_1
libblas=3.9.0=16_win64_mkl
libbrotlicommon=1.0.9=hcfcfb64_8
libbrotlidec=1.0.9=hcfcfb64_8
libbrotlienc=1.0.9=hcfcfb64_8
libcblas=3.9.0=16_win64_mkl
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.2.106=0
libcurand-dev=10.3.2.106=0
libcurl=8.1.1=h68f0423_0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=hcfcfb64_0
libexpat=2.5.0=h63175ca_1
libffi=3.4.2=h8ffe710_5
libgdal=3.6.2=h221c0cb_9
libglib=2.76.3=he8f3873_0
libhwloc=2.9.1=h51c2c0f_0
libiconv=1.17=h8ffe710_0
libjpeg-turbo=2.1.4=hcfcfb64_0
libkml=1.3.0=hf2ab4e4_1015
liblapack=3.9.0=16_win64_mkl
liblapacke=3.9.0=16_win64_mkl
libnetcdf=4.9.1=nompi_hc41bf00_101
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.39=h19919ed_0
libpq=15.3=ha9684e8_0
librttopo=1.1.0=he22b5cd_12
libspatialite=5.0.1=hfdcade0_23
libsqlite=3.42.0=hcfcfb64_0
libssh2=1.10.0=h9a1e1f7_3
libtiff=4.5.0=hf8721a0_2
libuv=1.44.2=h8ffe710_0
libwebp-base=1.3.0=hcfcfb64_0
libxcb=1.13=hcd874cb_1004
libxml2=2.10.4=hc3477c8_0
libzip=1.9.2=h519de47_1
libzlib=1.2.13=hcfcfb64_4
libzopfli=1.0.3=h0e60522_0
locket=1.0.0=pyhd8ed1ab_0
lz4-c=1.9.4=h2bbff1b_0
lzo=2.10=he774522_1000
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.2=py311ha68e1ae_0
matplotlib-base=3.7.1=py311h6e989c2_0
mkl=2022.1.0=h6a75c08_874
mkl-devel=2022.1.0=h57928b3_875
mkl-include=2022.1.0=h6a75c08_874
mpmath=1.3.0=pyhd8ed1ab_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=py_0
networkx=3.1=pyhd8ed1ab_0
numpy=1.24.3=py311h0b4df5a_0
openjpeg=2.5.0=ha2aaf27_2
openssl=3.1.0=hcfcfb64_3
ordered-set=4.1.0=pypi_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.1=py311hf63dbb6_1
partd=1.4.0=pyhd8ed1ab_0
patsy=0.5.3=pyhd8ed1ab_0
pcre2=10.40=h17e33f8_0
pillow=9.4.0=py311h76d9071_1
pip=23.1.2=pyhd8ed1ab_0
pixman=0.40.0=h8ffe710_0
platformdirs=3.5.1=pyhd8ed1ab_0
pooch=1.7.0=pyha770c72_3
poppler=23.03.0=h183ae7b_0
poppler-data=0.4.12=hd8ed1ab_0
postgresql=15.3=hd87cd2b_0
proj=9.1.1=heca977f_2
pthread-stubs=0.3=h3c9f919_1
pthreads-win32=2.9.1=hfa6e2cd_3
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=pyh0701188_6
python=3.11.3=h2628c8c_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.11=3_cp311
pytorch=2.0.1=py3.11_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_5
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.4.1=py311h59ca53f_0
pyyaml=6.0=py311ha68e1ae_5
rasterio=1.3.6=py311hc06ee40_0
requests=2.31.0=pyhd8ed1ab_0
scikit-image=0.20.0=py311hf63dbb6_1
scikit-learn=1.2.2=py311h142b183_2
scipy=1.10.1=py311h37ff6ca_3
seaborn=0.12.2=hd8ed1ab_0
seaborn-base=0.12.2=pyhd8ed1ab_0
setuptools=67.7.2=pyhd8ed1ab_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.10=hfb803bf_0
snuggs=1.4.7=py_0
sqlite=3.42.0=hcfcfb64_0
statsmodels=0.14.0=py311h59ca53f_1
sympy=1.12=pyh04b8f61_3
tbb=2021.9.0=h91493d7_0
threadpoolctl=3.1.0=pyh8a188c0_0
tifffile=2023.4.12=pyhd8ed1ab_0
tiledb=2.13.2=h3132609_0
tk=8.6.12=h8ffe710_0
toolz=0.12.0=pyhd8ed1ab_0
torchaudio=2.0.2=pypi_0
torchmeta=1.8.0=pypi_0
torchvision=0.15.2=pypi_0
tqdm=4.65.0=pyhd8ed1ab_1
typing-extensions=4.6.2=hd8ed1ab_0
typing_extensions=4.6.2=pyha770c72_0
tzdata=2023c=h71feb2d_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.0.2=pyhd8ed1ab_0
vc=14.3=hb25d44b_16
vc14_runtime=14.34.31931=h5081d32_16
vs2015_runtime=14.34.31931=hed1258a_16
wheel=0.40.0=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
xerces-c=3.2.4=h63175ca_2
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
yaml=0.2.5=h8ffe710_2
zfp=1.0.0=h63175ca_3
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=hcfcfb64_4
zlib-ng=2.0.7=hcfcfb64_0
zstd=1.5.2=h19a0ad4_0
